deploy/bot/src/requirements.txt
aiogram
asyncpg
aiohttp
python-dateutil
tzdata
timezonefinder
uvloop
python-dotenv
